refs #320 fixed regression test
git-svn-id: https://devel.codership.com/svn/galera/branches/0.7@1701 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ tests/regressions/lp578828/run.sh
@@ -1,4 +1,4 @@
-#!/bin/bash -eu
+#!/bin/bash -u
 
 declare -r DIST_BASE=$(cd $(dirname $0)/../..; pwd -P)
 TEST_BASE=${TEST_BASE:-"$DIST_BASE"}
@@ -26,6 +26,12 @@ $MYSQL -e "DROP TABLE IF EXISTS test.lp578828"
 $MYSQL -e "$create_stmt"
 $MYSQL -e "$insert_stmt"
 $MYSQL -e "$insert_stmt"
+if test $? == 0
+then
+    echo "duplicate insert succeeded"
+    exit 1
+fi
+
 
 $MYSQL -e "DROP TABLE test.lp578828"
